+++ src/main/java/tk/cth451/transitrailmod/blocks/TurnstileBlock.java
@@ -11,16 +11,22 @@ import net.minecraft.block.state.BlockState;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.BlockPos;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumWorldBlockLayer;
+import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import tk.cth451.transitrailmod.TransitRailMod;
+import tk.cth451.transitrailmod.TransitRailTab;
 import tk.cth451.transitrailmod.blocks.prototype.CustomDirectionBlock;
 import tk.cth451.transitrailmod.enums.EnumPassingDirection;
+import tk.cth451.transitrailmod.init.ModItems;
 
 public class TurnstileBlock extends CustomDirectionBlock{
 	
@@ -31,6 +37,7 @@ public class TurnstileBlock extends CustomDirectionBlock{
 	public TurnstileBlock(Material materialIn) {
 		super(Material.iron);
 		this.setUnlocalizedName("turnstile_block");
+		this.setCreativeTab(TransitRailMod.tabTransitRail);
 		this.setTickRandomly(true);
 		this.setDefaultState(this.getDefaultState()
 				.withProperty(ACTIVE, false)
@@ -100,7 +107,9 @@ public class TurnstileBlock extends CustomDirectionBlock{
 	
 	@Override
 	public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
-		worldIn.setBlockState(pos, state.withProperty(ACTIVE, false));
+		if ((Boolean) state.getValue(ACTIVE)){
+			worldIn.setBlockState(pos, state.withProperty(ACTIVE, false));
+		}
 	}
 	
 	// Appearance