+++ src/main/java/tk/cth451/transitrailmod/init/ModItems.java
@@ -12,7 +12,6 @@ import tk.cth451.transitrailmod.items.PlatformPanelItem;
 import tk.cth451.transitrailmod.items.StyleChanger;
 import tk.cth451.transitrailmod.items.TrainTicket;
 import tk.cth451.transitrailmod.items.TicketMachineItem;
-import tk.cth451.transitrailmod.items.TurnstileItem;
 
 public class ModItems {
 	public static Item style_changer;
@@ -20,19 +19,8 @@ public class ModItems {
 	public static Item closed_platform_panel_item;
 	public static Item platform_gate_item;
 	public static Item platform_panel_item;
-	public static Item turnstile_item;
 	public static Item train_ticket;
 	public static Item ticket_machine_item;
-	public static Item[] listOfItem = {
-		style_changer,
-		closed_platform_door_item,
-		closed_platform_panel_item,
-		platform_gate_item,
-		platform_panel_item,
-		turnstile_item,
-		train_ticket,
-		ticket_machine_item
-	};
 	
 	public static void init(){
 		style_changer = new StyleChanger();
@@ -40,7 +28,6 @@ public class ModItems {
 		closed_platform_panel_item = new ClosedPlatformPanelItem();
 		platform_gate_item = new PlatformGateItem();
 		platform_panel_item = new PlatformPanelItem();
-		turnstile_item = new TurnstileItem();
 		train_ticket = new TrainTicket();
 		ticket_machine_item = new TicketMachineItem();
 	}
@@ -51,7 +38,6 @@ public class ModItems {
 		GameRegistry.registerItem(closed_platform_panel_item, closed_platform_panel_item.getUnlocalizedName().substring(5)); 
 		GameRegistry.registerItem(platform_gate_item, platform_gate_item.getUnlocalizedName().substring(5)); 
 		GameRegistry.registerItem(platform_panel_item, platform_panel_item.getUnlocalizedName().substring(5)); 
-		GameRegistry.registerItem(turnstile_item, turnstile_item.getUnlocalizedName().substring(5)); 
 		GameRegistry.registerItem(train_ticket, train_ticket.getUnlocalizedName().substring(5)); 
 		GameRegistry.registerItem(ticket_machine_item, ticket_machine_item.getUnlocalizedName().substring(5)); 
 	}
@@ -62,7 +48,6 @@ public class ModItems {
 		registerRender(closed_platform_panel_item);
 		registerRender(platform_gate_item);
 		registerRender(platform_panel_item);
-		registerRender(turnstile_item);
 		registerRender(train_ticket);
 		registerRender(ticket_machine_item);
 	}